Core: Split mux.c POSIX/Windows implementations
